Move RDO check to extension
