+++ src/Frontend/Dfe.Complete/Models/TabNavigationModel.cs
@@ -1,4 +1,5 @@
 ﻿using Dfe.Complete.Domain.Enums;
+using Dfe.Complete.Extensions;
 
 namespace Dfe.Complete.Models;
 
@@ -13,7 +14,7 @@ public class TabNavigationModel(string currentTab)
 
     public static bool UserHasTabAccess(ProjectTeam userTeam, string tabName)
     {
-        if (tabName == YourTeamProjectsTabName) return EnumHelper.TeamIsRdo(userTeam) || userTeam == ProjectTeam.RegionalCaseWorkerServices;
+        if (tabName == YourTeamProjectsTabName) return userTeam.TeamIsRdo() || userTeam == ProjectTeam.RegionalCaseWorkerServices;
         return true;
     }
 }
